backend/requirements.txt
tensorflow-cpu==2.20.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==10.4.0
numpy>=2.1.0
python-multipart==0.0.12
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.1
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
